[services] Tunnelbroker - Fix `DeliveryBrokerTest.cpp` warnings while building by Clang
Summary:
Fixes build types conversion warnings while building by Clang:

```
DeliveryBrokerTest.cpp:50:22: non-constant-expression cannot be narrowed from type 'time_t' (aka 'long') to 'uint64_t' (aka 'unsigned long
long') in initializer list [-Wc++11-narrowing]
[build]       .deliveryTag = std::time(0),
```

Test Plan: Run `yarn test-tunnelbroker-service` to build and run the tests.

Reviewers: karol-bisztyga, ashoat

Reviewed By: ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3882

+++ services/tunnelbroker/test/DeliveryBrokerTest.cpp
@@ -47,7 +47,7 @@ TEST(DeliveryBrokerTest, CheckPushAndPopOnGeneratedValues) {
       "mobile:" + tools::generateRandomString(DEVICEID_CHAR_LENGTH);
   const DeliveryBrokerMessage message{
       .messageID = tools::generateUUID(),
-      .deliveryTag = std::time(0),
+      .deliveryTag = static_cast<uint64_t>(std::time(0)),
       .fromDeviceID =
           "mobile:" + tools::generateRandomString(DEVICEID_CHAR_LENGTH),
       .payload = tools::generateRandomString(512)};
